[ENH] Refactoring of nipype.interfaces.utility
Split utility.py into a module
