+++ nipype/interfaces/utility/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+"""
+Package contains interfaces for using existing functionality in other packages
+
+Requires Packages to be installed
+"""
+from __future__ import print_function, division, unicode_literals, absolute_import
+__docformat__ = 'restructuredtext'
+
+from .base import (IdentityInterface, Rename, Select, Split, Merge,
+                   AssertEqual)
+from .csv import CSVReader
+from .wrappers import Function
